backend/x_post/requirements.txt
requests
python-dotenv
tweepy
requests_oauthlib
